Part 4D done

+++ exercises/part4/blog/controllers/blog.js
@@ -1,15 +1,33 @@
 const blogRouter = require('express').Router()
 const Blog = require('../models/blog')
+const User = require('../models/user')
+const jwt = require('jsonwebtoken')
 
-blogRouter.post('/', async (request, response,next) => {
-  const blog = new Blog(request.body)
+
+blogRouter.post('/', async (request, response, next) => {
+  const {title, author, url, likes} = request.body
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+  if (!decodedToken.id) {
+    return response.status(401).json( { error: 'token invalid' })
+  }
+
+  const blogAuthor = request.user
+  const blog = new Blog({
+    title,
+    author,
+    url,
+    likes,
+    creator: blogAuthor._id
+  })
   const savedBlog = await blog.save({ runValidators: true })
+  blogAuthor.blogs = blogAuthor.blogs.concat(savedBlog._id)
+  await blogAuthor.save()
   response.status(201).json(savedBlog)
 
 })
 
 blogRouter.get('/', async (request, response) => {
-  const blogs = await Blog.find({})
+  const blogs = await Blog.find({}).populate('creator', { username: 1, name: 1})
   response.json(blogs)
 })
 
@@ -19,6 +37,19 @@ blogRouter.get('/:id', async (request, response) => {
 })
 
 blogRouter.delete('/:id', async (request, response) => {
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+
+  if (!decodedToken.id) {
+    return response.status(401).json( { error: 'token invalid' })
+  }
+  const blogToDelete = await Blog.findById(request.params.id)
+
+  if (blogToDelete){
+    if (blogToDelete.creator.toString() !== decodedToken.id) {
+      return response.status(401).json({ error: 'invalid user' })
+    }
+  }
+
   await Blog.findByIdAndDelete(request.params.id)
   response.status(204).end()
 })
